# src/main/resources/openapi/Device-Manager-v1.0.0.yaml
---
openapi: 3.0.1
info:
  title: Device Manager
  description: This is Swagger UI environment generated for the project Device Manager.
  version: '1.0'
servers:
  - url: https://serverRoot/api/device-manager
paths:
  "/v1/device":
    get:
      tags:
        - Device
      summary: Fetch devices.
      operationId: v1FetchDevices
      parameters:
        - name: brand
          in: query
          required: false
          schema:
            type: string
          description: Filter devices by brand.
        - name: state
          in: query
          required: false
          schema:
            type: string
            enum:
              - available
              - in-use
              - inactive
          description: Filter devices by state.
      responses:
        '200':
          description: Ok
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/DeviceResponseDto'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/Error"
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/Error"
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/Error"
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/Error"
        '405':
          description: Method Not allowed
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/Error"
        '409':
          description: Conflict
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/Error"
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/Error"
    post:
      tags:
        - Device
      summary: Create a device.
      operationId: v1CreateDevice
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DeviceRequestDto'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/DeviceResponseDto"
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/Error"
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/Error"
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/Error"
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/Error"
        '405':
          description: Method Not allowed
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/Error"
        '409':
          description: Conflict
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/Error"
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/Error"
  "/v1/device/{id}":
    get:
      tags:
        - Device
      summary: Fetch a device by ID.
      operationId: v1FetchDeviceById
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          description: Unique identifier for the device.
      responses:
        '200':
          description: Ok
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeviceResponseDto'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/Error"
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/Error"
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/Error"
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/Error"
        '405':
          description: Method Not allowed
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/Error"
        '409':
          description: Conflict
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/Error"
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/Error"
    patch:
      tags:
        - Device
      summary: Update a device by ID
      operationId: v1UpdateDeviceById
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          description: Unique identifier for the device.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DeviceRequestDto'
      responses:
        '200':
          description: Ok
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeviceResponseDto'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/Error"
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/Error"
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/Error"
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/Error"
        '405':
          description: Method Not allowed
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/Error"
        '409':
          description: Conflict
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/Error"
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/Error"
    delete:
      tags:
        - Device
      summary: Delete a device.
      operationId: v1DeleteDevice
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          description: Unique identifier for the device.
      responses:
        '204':
          description: No Content
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/Error"
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/Error"
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/Error"
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/Error"
        '405':
          description: Method Not allowed
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/Error"
        '409':
          description: Conflict
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/Error"
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/Error"
components:
  schemas:
    DeviceResponseDto:
      type: object
      properties:
        id:
          type: string
          description: "Unique identifier for the device."
        name:
          type: string
          description: "Name of the device."
        brand:
          type: string
          description: "Brand of the device."
        state:
          $ref: '#/definitions/DeviceState'
        creationTime:
          type: string
          format: date-time
          description: "Timestamp when the device was created."
    DeviceRequestDto:
      type: object
      properties:
        name:
          type: string
          description: "Name of the device."
        brand:
          type: string
          description: "Brand of the device."
        state:
          type: string
          description: "State of the device."
    DeviceState:
      type: string
      enum:
        - available
        - in-use
        - inactive
      description: "State of the device."
    Error:
      required:
        - code
        - reason
      type: object
      properties:
        code:
          type: string
          description: Application related code.
        reason:
          type: string
          description: Text that explains the reason for error.
        message:
          type: string
          description: "(optional) Text that provide more details and corrective actions
            related to the error."
        status:
          type: integer
          description: "(optional) http error code extension like 400-2"